Make changes to reduce code duplications in WebSocket implementation
