+++ components/org.wso2.transport.http.netty/src/main/java/org/wso2/transport/http/netty/contractimpl/websocket/DefaultWebSocketConnectorFuture.java
@@ -1,0 +1,100 @@
+/*
+ *  Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+
+package org.wso2.transport.http.netty.contractimpl.websocket;
+
+import org.wso2.transport.http.netty.contract.websocket.WebSocketBinaryMessage;
+import org.wso2.transport.http.netty.contract.websocket.WebSocketCloseMessage;
+import org.wso2.transport.http.netty.contract.websocket.WebSocketConnection;
+import org.wso2.transport.http.netty.contract.websocket.WebSocketConnectorException;
+import org.wso2.transport.http.netty.contract.websocket.WebSocketConnectorFuture;
+import org.wso2.transport.http.netty.contract.websocket.WebSocketConnectorListener;
+import org.wso2.transport.http.netty.contract.websocket.WebSocketControlMessage;
+import org.wso2.transport.http.netty.contract.websocket.WebSocketInitMessage;
+import org.wso2.transport.http.netty.contract.websocket.WebSocketTextMessage;
+
+/**
+ * Default implementation of {@link WebSocketConnectorFuture}.
+ */
+public class DefaultWebSocketConnectorFuture implements WebSocketConnectorFuture {
+
+    private WebSocketConnectorListener wsConnectorListener;
+
+
+    @Override
+    public void setWSConnectorListener(WebSocketConnectorListener wsConnectorListener) {
+        this.wsConnectorListener = wsConnectorListener;
+    }
+
+    @Override
+    public void notifyWSListener(WebSocketInitMessage initMessage) throws WebSocketConnectorException {
+        if (wsConnectorListener == null) {
+            throw new WebSocketConnectorException("WebSocket connector listener is not set");
+        }
+        wsConnectorListener.onMessage(initMessage);
+    }
+
+    @Override
+    public void notifyWSListener(WebSocketTextMessage textMessage) throws WebSocketConnectorException {
+        if (wsConnectorListener == null) {
+            throw new WebSocketConnectorException("WebSocket connector listener is not set");
+        }
+        wsConnectorListener.onMessage(textMessage);
+    }
+
+    @Override
+    public void notifyWSListener(WebSocketBinaryMessage binaryMessage) throws WebSocketConnectorException {
+        if (wsConnectorListener == null) {
+            throw new WebSocketConnectorException("WebSocket connector listener is not set");
+        }
+        wsConnectorListener.onMessage(binaryMessage);
+    }
+
+    @Override
+    public void notifyWSListener(WebSocketControlMessage controlMessage) throws WebSocketConnectorException {
+        if (wsConnectorListener == null) {
+            throw new WebSocketConnectorException("WebSocket connector listener is not set");
+        }
+        wsConnectorListener.onMessage(controlMessage);
+    }
+
+    @Override
+    public void notifyWSListener(WebSocketCloseMessage closeMessage) throws WebSocketConnectorException {
+        if (wsConnectorListener == null) {
+            throw new WebSocketConnectorException("WebSocket connector listener is not set");
+        }
+        wsConnectorListener.onMessage(closeMessage);
+    }
+
+    @Override
+    public void notifyWSListener(WebSocketConnection webSocketConnection, Throwable throwable)
+            throws WebSocketConnectorException {
+        if (wsConnectorListener == null) {
+            throw new WebSocketConnectorException("WebSocket connector listener is not set");
+        }
+        wsConnectorListener.onError(webSocketConnection, throwable);
+    }
+
+    @Override
+    public void notifyWSIdleTimeout(WebSocketControlMessage controlMessage) throws WebSocketConnectorException {
+        if (wsConnectorListener == null) {
+            throw new WebSocketConnectorException("WebSocket connector listener is not set");
+        }
+        wsConnectorListener.onIdleTimeout(controlMessage);
+    }
+}
