+++ components/org.wso2.transport.http.netty/src/test/java/org/wso2/transport/http/netty/websocket/client/WebSocketClientFunctionalityTestCase.java
@@ -248,7 +248,9 @@ public class WebSocketClientFunctionalityTestCase {
     }
 
     private ClientHandshakeFuture handshake(WebSocketConnectorListener connectorListener) {
-        return clientConnector.connect(connectorListener);
+        ClientHandshakeFuture clientHandshakeFuture = clientConnector.connect();
+        clientHandshakeFuture.setWSConnectorListener(connectorListener);
+        return clientHandshakeFuture;
     }
 
     private WebSocketConnection getWebSocketConnectionSync(WebSocketConnectorListener connectorListener)
