+++ components/org.wso2.transport.http.netty/src/test/java/org/wso2/transport/http/netty/websocket/client/WebSocketClientHandshakeFunctionalityTestCase.java
@@ -232,6 +232,8 @@ public class WebSocketClientHandshakeFunctionalityTestCase {
     }
 
     private ClientHandshakeFuture handshake(WebSocketConnectorListener connectorListener) {
-        return clientConnector.connect(connectorListener);
+        ClientHandshakeFuture clientHandshakeFuture = clientConnector.connect();
+        clientHandshakeFuture.setWSConnectorListener(connectorListener);
+        return clientHandshakeFuture;
     }
 }
