+++ components/org.wso2.transport.http.netty/src/test/java/org/wso2/transport/http/netty/websocket/server/WebSocketServerHandshakeFunctionalityTestCase.java
@@ -46,7 +46,7 @@ import java.util.concurrent.TimeUnit;
 public class WebSocketServerHandshakeFunctionalityTestCase {
 
     private static final int countdownLatchTimeout = 10;
-        DefaultHttpWsConnectorFactory httpConnectorFactory;
+    private DefaultHttpWsConnectorFactory httpConnectorFactory;
     private ServerConnector serverConnector;
     private WebSocketServerHandshakeFunctionalityListener listener;
